requirements.txt
pip
torch
langchain
unstructured
langchain-community
langchain-groq
unstructured
#git+https://github.com/facebookresearch/detectron2.git
layoutparser[layoutmodels,tesseract]
pytesseract
Pillow==9.0.0
pybind11
chromadb
Cython
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
#streamlit-extras
python-dotenv


